Move new nodes port object
#559
