+++ de.bund.bfr.knime.fsklab.nodes/plugin.xml
@@ -8,20 +8,20 @@
       color="000000"
       hidden="false"
       name="FSK object"
-      objectClass="de.bund.bfr.knime.fsklab.FskPortObject"
-      objectSerializer="de.bund.bfr.knime.fsklab.FskPortObject$Serializer"
-      specClass="de.bund.bfr.knime.fsklab.FskPortObjectSpec"
-      specSerializer="de.bund.bfr.knime.fsklab.FskPortObjectSpec$Serializer">
+      objectClass="de.bund.bfr.knime.fsklab.v1_9.FskPortObject"
+      objectSerializer="de.bund.bfr.knime.fsklab.v1_9.FskPortObject$Serializer"
+      specClass="de.bund.bfr.knime.fsklab.v1_9.FskPortObjectSpec"
+      specSerializer="de.bund.bfr.knime.fsklab.v1_9.FskPortObjectSpec$Serializer">
     </portType>
     
     <portType
       color="FF4500"
       hidden="false"
       name="Combined FSK object"
-      objectClass="de.bund.bfr.knime.fsklab.CombinedFskPortObject"
-      objectSerializer="de.bund.bfr.knime.fsklab.CombinedFskPortObject$Serializer"
-      specClass="de.bund.bfr.knime.fsklab.CombinedFskPortObjectSpec"
-      specSerializer="de.bund.bfr.knime.fsklab.CombinedFskPortObjectSpec$Serializer">
+      objectClass="de.bund.bfr.knime.fsklab.v1_9.CombinedFskPortObject"
+      objectSerializer="de.bund.bfr.knime.fsklab.v1_9.CombinedFskPortObject$Serializer"
+      specClass="de.bund.bfr.knime.fsklab.v1_9.CombinedFskPortObjectSpec"
+      specSerializer="de.bund.bfr.knime.fsklab.v1_9.CombinedFskPortObjectSpec$Serializer">
     </portType>
   </extension>
